forgot to fix warnings

+++ StarEngine/jni/Components/SpriteComponent.cpp
@@ -44,14 +44,14 @@ namespace star
 
 	void SpriteComponent::CreateSquare()
 	{
-		m_Vertices[0] = m_Width;
-		m_Vertices[1] = m_Heigth;
+		m_Vertices[0] = (GLfloat)m_Width;
+		m_Vertices[1] = (GLfloat)m_Heigth;
 		m_Vertices[2] = 0;
-		m_Vertices[3] = m_Width;
+		m_Vertices[3] = (GLfloat)m_Width;
 		m_Vertices[4] = 0;
 		m_Vertices[5] = 0;
 		m_Vertices[6] = 0;
-		m_Vertices[7] = m_Heigth;
+		m_Vertices[7] = (GLfloat)m_Heigth;
 		m_Vertices[8] = 0;
 		m_Vertices[9] = 0;
 		m_Vertices[10] = 0;
